Корректировка метода подписки/отписки

+++ backend/users/views.py
@@ -70,34 +70,35 @@ class UserViewSet(viewsets.ModelViewSet):
                                             context={'request': request})
         return self.get_paginated_response(serializer.data)
 
-    @action(detail=True, methods=['post'],
+    @action(detail=True, methods=['post', 'delete'],
             permission_classes=(IsAuthenticated,))
     def subscribe(self, request, **kwargs):
         author = get_object_or_404(User, id=kwargs['pk'])
-        serializer = SubscriptionActionSerializer(
-            data={'author': author}, context={'request': request})
-        serializer.is_valid(raise_exception=True)
-        subscription, created = Subscription.objects.get_or_create(
-            user=request.user, author=author)
-        if not created:
-            return Response(
-                {'error': 'Вы уже подписаны на этого пользователя',
-                 'is_subscribed': True},
-                status=status.HTTP_400_BAD_REQUEST)
-        response_data = serializer.data
-        response_data['message'] = 'Подписка успешно создана'
-        return Response(response_data, status=status.HTTP_201_CREATED)
 
-    @subscribe.mapping.delete
-    def delete_subscribe(self, request, **kwargs):
-        author = get_object_or_404(User, id=kwargs['pk'])
-        subscription = Subscription.objects.filter(
-            user=request.user, author=author).first()
-        if not subscription:
-            return Response({'errors': 'Подписка не существует',
-                             'is_subscribed': False},
-                            status=status.HTTP_400_BAD_REQUEST)
-        subscription.delete()
-        response_data = {'message': 'Подписка успешно удалена',
-                         'is_subscribed': False}
-        return Response(response_data, status=status.HTTP_204_NO_CONTENT)
+        if request.method == 'POST':
+            serializer = SubscriptionActionSerializer(
+                data={'author': author}, context={'request': request})
+            serializer.is_valid(raise_exception=True)
+            subscription, created = Subscription.objects.get_or_create(
+                user=request.user, author=author)
+            if not created:
+                return Response(
+                    {'error': 'Вы уже подписаны на этого пользователя',
+                     'is_subscribed': True},
+                    status=status.HTTP_400_BAD_REQUEST)
+            response_data = serializer.data
+            response_data['message'] = 'Подписка успешно создана'
+            return Response(response_data, status=status.HTTP_201_CREATED)
+
+        if request.method == 'DELETE':
+            subscription = Subscription.objects.filter(
+                user=request.user, author=author).first()
+            if not subscription:
+                return Response(
+                    {'errors': 'Подписка не существует',
+                     'is_subscribed': False},
+                    status=status.HTTP_400_BAD_REQUEST)
+            subscription.delete()
+            response_data = {'message': 'Подписка успешно удалена',
+                             'is_subscribed': False}
+            return Response(response_data, status=status.HTTP_204_NO_CONTENT)
